Move this test to the right class
